@robehn 2023-10-27
Looks good, but these numbers:
```
LongDivMod.testRemainderUnsigned                 1024       negative  avgt   10     7733.038 ±  3.856  ns/op
IntegerDivMod.testRemainderUnsigned             1024       negative  avgt   10  30159.407 ±  6.716  ns/op

```
What up here ? How come in before case Long is 4-5x faster than Integer?

I don't see an explanation in other PR either ?