@Hamlin-Li 2023-10-27
Good find.

Let me try to explain.

For l.l.Long, remainderUnsigned has a quick path for negative divisor (This was discussed in https://github.com/openjdk/jdk/pull/16346), it's much quicker than the case of positive divisor.

For j.l.Integer
```
    public static int remainderUnsigned(int dividend, int divisor) {
        // In lieu of tricky code, for now just use long arithmetic.
        return (int)(toUnsignedLong(dividend) % toUnsignedLong(divisor));
    }
```
it will call Long's rem, and the value of divisor passed to Long rem will be positive in the sense of Long.
To verify this, you can compare the Long's rem data before the patch, positive one is much lower than negative one.
```
LongDivMod.testRemainderUnsigned                 1024       positive  avgt   10  28841.429 ±  6.294  ns/op
LongDivMod.testRemainderUnsigned                 1024       negative  avgt   10   7733.038 ±  3.856  ns/op
```

Hope this answers your question.