@VladimirKempik 2023-10-27
So, you win 45% on
LongDivMod.testRemainderUnsigned                 1024       positive

but loose almost 4x on
LongDivMod.testRemainderUnsigned                 1024       negative

Does this intrinsic ( for LONGs) make sense at all? or better leave it as is ?